[hotfix] Various code cleanup in serializers

+++ flink-core/src/main/java/org/apache/flink/api/common/typeutils/base/ByteSerializer.java
@@ -18,20 +18,24 @@
 
 package org.apache.flink.api.common.typeutils.base;
 
-import java.io.IOException;
-
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.core.memory.DataInputView;
 import org.apache.flink.core.memory.DataOutputView;
 
+import java.io.IOException;
+
+/**
+ * Type serializer for {@code Byte}.
+ */
 @Internal
 public final class ByteSerializer extends TypeSerializerSingleton<Byte> {
 
 	private static final long serialVersionUID = 1L;
-	
+
+	/** Sharable instance of the ByteSerializer. */
 	public static final ByteSerializer INSTANCE = new ByteSerializer();
-	
-	private static final Byte ZERO = Byte.valueOf((byte) 0);
+
+	private static final Byte ZERO = (byte) 0;
 
 	@Override
 	public boolean isImmutableType() {
@@ -47,7 +51,7 @@ public final class ByteSerializer extends TypeSerializerSingleton<Byte> {
 	public Byte copy(Byte from) {
 		return from;
 	}
-	
+
 	@Override
 	public Byte copy(Byte from, Byte reuse) {
 		return from;
@@ -60,14 +64,14 @@ public final class ByteSerializer extends TypeSerializerSingleton<Byte> {
 
 	@Override
 	public void serialize(Byte record, DataOutputView target) throws IOException {
-		target.writeByte(record.byteValue());
+		target.writeByte(record);
 	}
 
 	@Override
 	public Byte deserialize(DataInputView source) throws IOException {
-		return Byte.valueOf(source.readByte());
+		return source.readByte();
 	}
-	
+
 	@Override
 	public Byte deserialize(Byte reuse, DataInputView source) throws IOException {
 		return deserialize(source);
