+++ flink-core/src/main/java/org/apache/flink/api/common/typeutils/base/DoubleSerializer.java
@@ -18,20 +18,24 @@
 
 package org.apache.flink.api.common.typeutils.base;
 
-import java.io.IOException;
-
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.core.memory.DataInputView;
 import org.apache.flink.core.memory.DataOutputView;
 
+import java.io.IOException;
+
+/**
+ * Type serializer for {@code Double}.
+ */
 @Internal
 public final class DoubleSerializer extends TypeSerializerSingleton<Double> {
 
 	private static final long serialVersionUID = 1L;
-	
+
+	/** Sharable instance of the DoubleSerializer. */
 	public static final DoubleSerializer INSTANCE = new DoubleSerializer();
-	
-	private static final Double ZERO = Double.valueOf(0);
+
+	private static final Double ZERO = 0.0;
 
 	@Override
 	public boolean isImmutableType() {
@@ -47,7 +51,7 @@ public final class DoubleSerializer extends TypeSerializerSingleton<Double> {
 	public Double copy(Double from) {
 		return from;
 	}
-	
+
 	@Override
 	public Double copy(Double from, Double reuse) {
 		return from;
@@ -60,14 +64,14 @@ public final class DoubleSerializer extends TypeSerializerSingleton<Double> {
 
 	@Override
 	public void serialize(Double record, DataOutputView target) throws IOException {
-		target.writeDouble(record.doubleValue());
+		target.writeDouble(record);
 	}
 
 	@Override
 	public Double deserialize(DataInputView source) throws IOException {
-		return Double.valueOf(source.readDouble());
+		return source.readDouble();
 	}
-	
+
 	@Override
 	public Double deserialize(Double reuse, DataInputView source) throws IOException {
 		return deserialize(source);
