+++ flink-core/src/main/java/org/apache/flink/api/common/typeutils/base/IntSerializer.java
@@ -18,20 +18,24 @@
 
 package org.apache.flink.api.common.typeutils.base;
 
-import java.io.IOException;
-
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.core.memory.DataInputView;
 import org.apache.flink.core.memory.DataOutputView;
 
+import java.io.IOException;
+
+/**
+ * Type serializer for {@code Integer} (and {@code int}, via auto-boxing).
+ */
 @Internal
 public final class IntSerializer extends TypeSerializerSingleton<Integer> {
 
 	private static final long serialVersionUID = 1L;
-	
+
+	/** Sharable instance of the IntSerializer. */
 	public static final IntSerializer INSTANCE = new IntSerializer();
-	
-	private static final Integer ZERO = Integer.valueOf(0);
+
+	private static final Integer ZERO = 0;
 
 	@Override
 	public boolean isImmutableType() {
@@ -47,7 +51,7 @@ public final class IntSerializer extends TypeSerializerSingleton<Integer> {
 	public Integer copy(Integer from) {
 		return from;
 	}
-	
+
 	@Override
 	public Integer copy(Integer from, Integer reuse) {
 		return from;
@@ -60,14 +64,14 @@ public final class IntSerializer extends TypeSerializerSingleton<Integer> {
 
 	@Override
 	public void serialize(Integer record, DataOutputView target) throws IOException {
-		target.writeInt(record.intValue());
+		target.writeInt(record);
 	}
 
 	@Override
 	public Integer deserialize(DataInputView source) throws IOException {
-		return Integer.valueOf(source.readInt());
+		return source.readInt();
 	}
-	
+
 	@Override
 	public Integer deserialize(Integer reuse, DataInputView source) throws IOException {
 		return deserialize(source);
